add: admin pet ep

+++ src/main/java/com/example/pupilmed/repositories/OwnerRepository.java
@@ -1,13 +1,12 @@
 package com.example.pupilmed.repositories;
 
-import com.example.pupilmed.models.database.owner.Owner;
-import com.example.pupilmed.models.database.user.User;
-import com.example.pupilmed.models.database.vet.Vet;
+import com.example.pupilmed.models.database.Owner;
+import com.example.pupilmed.models.database.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OwnerRepository extends JpaRepository<Owner,Long> {
+public interface OwnerRepository extends JpaRepository<Owner,Integer> {
 
     Owner findOwnerByUser(User user);
     Owner findOwnerByUser_Username(String username);
