+++ src/main/java/com/example/pupilmed/repositories/PetRepository.java
@@ -1,7 +1,8 @@
 package com.example.pupilmed.repositories;
 
-import com.example.pupilmed.models.database.owner.Owner;
-import com.example.pupilmed.models.database.pet.Pet;
+import com.example.pupilmed.models.database.Owner;
+import com.example.pupilmed.models.database.Pet;
+import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,12 +10,12 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PetRepository extends JpaRepository<Pet,Long> {
+public interface PetRepository extends JpaRepository<Pet,Integer> {
 
     List<Pet> findPetByOwner_Id(Integer id);
     List<Pet> findByOwner(Owner owner);
     Pet findPetByNameAndOwner(String name, Owner owner);
     boolean existsByNameAndOwner(String name, Owner owner);
     Optional<Pet> getPetById(Integer id);
-
+//    boolean existsById(Integer id);
 }
