+++ src/main/java/com/example/pupilmed/repositories/RecommendationRepository.java
@@ -1,11 +1,11 @@
 package com.example.pupilmed.repositories;
 
-import com.example.pupilmed.models.database.recommendation.Recommendation;
+import com.example.pupilmed.models.database.Recommendation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecommendationRepository extends JpaRepository<Recommendation,Long> {
+public interface RecommendationRepository extends JpaRepository<Recommendation,Integer> {
 
 //    Recommendation getRecommendationById(Integer id);
