+++ src/main/java/com/example/pupilmed/repositories/UserRepository.java
@@ -1,14 +1,14 @@
 package com.example.pupilmed.repositories;
 
-import com.example.pupilmed.models.database.user.Role;
-import com.example.pupilmed.models.database.user.User;
+import com.example.pupilmed.models.database.Role;
+import com.example.pupilmed.models.database.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Long> {
+public interface UserRepository extends JpaRepository<User,Integer> {
     Optional<User> findByUsername(String username);
 
     boolean existsByUsernameAndRole(String username, Role role);
